requirements.txt
fastapi==0.127.0
pydantic==2.12.5
python-dotenv==1.2.1
starlette==0.50.0
uvicorn==0.40.0
python-slugify==8.0.1